chore: move metrics
