+++ rust/numaflow-core/src/source.rs
@@ -1,2 +1,1 @@
 pub(crate) mod user_defined;
-
